api/requirements.txt
azure-functions
pytest

